Float32 avoid memoryerror

+++ hdf5_utils.py
@@ -398,8 +398,8 @@ class H52Npy:
                 tr = nonseis_grp[dset][:]
                 nonseis_trs.append(np.hstack([tr, label]))
 
-        seis_trs = np.asarray(seis_trs)
-        nonseis_trs = np.asarray(nonseis_trs)
+        seis_trs = np.asarray(seis_trs, dtype=np.float32)
+        nonseis_trs = np.asarray(nonseis_trs, dtype=np.float32)
 
         all_tr = np.vstack([seis_trs, nonseis_trs])
